USB: Change variable type
Make it consistent with the rest of the source code (MagiC
variable is defined as a short).

+++ sys/usb/src.km/tos-common/magic.c
@@ -66,7 +66,7 @@ typedef struct
 } MAGX_COOKIE;
 
 /* Are we running on MagiC? */
-int MagiC = -1;
+short MagiC = -1;
 
 
 /* Code taken from TOS.HYP
